[CHORE] 순서 관련 로직들 수정

+++ src/main/kotlin/com/official/memento/todo/domain/vo/ClaudeAiChatClientOutputPort.kt
@@ -5,7 +5,7 @@ import com.official.memento.todo.domain.entity.ToDo
 interface ClaudeAiChatClientOutputPort {
     fun prioritizeTodo(
             todoList: List<ToDo>,
-            orderList: List<Int>,
+            orderList: List<Double>,
             personalInfo: String
     ): List<PrioritizedToDo>
 }
